refactor(contribute): move to /project/[id]/contribute
