+++ src/lib/ContributeButton.svelte
@@ -1,11 +1,14 @@
 <script>
-  
   export let project;
-  
 </script>
 
-<a href="/contribute/{project.id}" class="flex items-center justify-center w-full mt-5 text-xs font-medium leading-loose border border-solid whitespace-nowrap rounded-[48px] bg-cyan-950 border-teal-950 text-zinc-100">
-  <span class="overflow-hidden gap-2 self-stretch px-7 py-3.5 my-auto rounded-3xl max-md:px-5 text-xl">
+<a
+  href="/project/{project.id}/contribute"
+  class="flex items-center justify-center w-full mt-5 text-xs font-medium leading-loose border border-solid whitespace-nowrap rounded-[48px] bg-cyan-950 border-teal-950 text-zinc-100"
+>
+  <span
+    class="overflow-hidden gap-2 self-stretch px-7 py-3.5 my-auto rounded-3xl max-md:px-5 text-xl"
+  >
     Contribute
   </span>
-</a> 
+</a>
